Support generic functions, inline task
When calling generic functions, we don't have (simple) access to the
parameter types that the function was actually instantiated with,
therefore the generated shim doesn't work.

Instead of using a separate `toTask` macro which requires constructing
an extra shim function, construct the closure environment directly in
`spawn` using a `tuple` which allows the compiler to deduce the correct
types.

Since we're here, fix a few other things:

* get rid of `Task` entirely, along with `toTask` which is quite buggy
* inline closure environment creation in `spawn` creating a tuple for it
in shared memory
* get rid of separate destructor - we can release memory in the
callback, reducing overhead further
* support move-only types in refc